Implement ezProcessGroup for non windows platforms.
